+++ android/app/src/main/java/com/yabrag/ecgeeegblemonitor/analysis/EegAnalysisPackage.kt
@@ -1,0 +1,17 @@
+package com.yabrag.ecgeeegblemonitor.analysis
+
+import com.facebook.react.ReactPackage
+import com.facebook.react.bridge.NativeModule
+import com.facebook.react.bridge.ReactApplicationContext
+import com.facebook.react.uimanager.ViewManager
+
+@Suppress("DEPRECATION")
+class EegAnalysisPackage : ReactPackage {
+  override fun createNativeModules(reactContext: ReactApplicationContext): List<NativeModule> {
+    return listOf(EegAnalysisModule(reactContext))
+  }
+
+  override fun createViewManagers(reactContext: ReactApplicationContext): List<ViewManager<*, *>> {
+    return emptyList()
+  }
+}